Changes to fix OSX errors
