+++ archimesh/src/achm_window_maker.py
@@ -28,13 +28,13 @@ import bpy
 import math
 # noinspection PyUnresolvedReferences
 from bpy.props import *
-from arch_tools import *
+from achm_tools import *
 
 
 # ------------------------------------------------------------------
 # Define operator class to create object
 # ------------------------------------------------------------------
-class WINDOWS(bpy.types.Operator):
+class AchmWindows(bpy.types.Operator):
     bl_idname = "mesh.archimesh_window"
     bl_label = "Rail Windows"
     bl_description = "Rail Windows Generator"
@@ -81,7 +81,7 @@ def create_object(self, context):
     mainobject.WindowObjectGenerator.add()
 
     # we shape the main object and create other objects as children
-    shape_mesh_and_create_children(mainobject)
+    shape_mesh_and_create_children(mainobject, mainmesh)
 
     # we select, and activate, main object
     mainobject.select = True
@@ -97,15 +97,12 @@ def create_object(self, context):
 def update_object(self, context):
     # When we update, the active object is the main object
     o = bpy.context.active_object
+    oldmesh = o.data
+    oldname = o.data.name
     # Now we deselect that object to not delete it.
     o.select = False
-    # Remove mesh data
-    o.data.user_clear()
-    bpy.data.meshes.remove(o.data)
     # and we create a new mesh
-    objmesh = bpy.data.meshes.new("WindowFrame")
-    o.data = objmesh
-    o.data.use_fake_user = True
+    tmp_mesh = bpy.data.meshes.new("temp")
     # deselect all objects
     for obj in bpy.data.objects:
         obj.select = False
@@ -125,10 +122,10 @@ def update_object(self, context):
                 # clear child data
                 child.hide = False  # must be visible to avoid bug
                 child.hide_render = False  # must be visible to avoid bug
-                child.data.user_clear()
-                bpy.data.meshes.remove(child.data)
+                old = child.data
                 child.select = True
                 bpy.ops.object.delete()
+                bpy.data.meshes.remove(old)
             except:
                 dummy = -1
 
@@ -145,7 +142,11 @@ def update_object(self, context):
     remove_children(o)
 
     # Finally we create all that again (except main object),
-    shape_mesh_and_create_children(o, True)
+    shape_mesh_and_create_children(o, tmp_mesh, True)
+    o.data = tmp_mesh
+    # Remove data (mesh of active object),
+    bpy.data.meshes.remove(oldmesh)
+    tmp_mesh.name = oldname
     # and select, and activate, the main object
     o.select = True
     bpy.context.scene.objects.active = o
@@ -157,13 +158,13 @@ def update_object(self, context):
 # And, for the others, it creates object and mesh.
 # ------------------------------------------------------------------------------
 # noinspection PyUnusedLocal
-def shape_mesh_and_create_children(mainobject, update=False):
+def shape_mesh_and_create_children(mainobject, tmp_mesh, update=False):
     mp = mainobject.WindowObjectGenerator[0]
     # Create only mesh, because the object is created before
     if mp.opentype == "1":
-        generate_rail_window(mainobject, mp, mainobject.data)
+        generate_rail_window(mainobject, mp, tmp_mesh)
     else:
-        generate_leaf_window(mainobject, mp, mainobject.data)
+        generate_leaf_window(mainobject, mp, tmp_mesh)
 
     remove_doubles(mainobject)
     set_normals(mainobject)
@@ -314,7 +315,7 @@ bpy.types.Object.WindowObjectGenerator = bpy.props.CollectionProperty(type=Objec
 # ------------------------------------------------------------------
 # Define panel class to modify object
 # ------------------------------------------------------------------
-class WindowObjectgeneratorpanel(bpy.types.Panel):
+class AchmWindowObjectgeneratorpanel(bpy.types.Panel):
     bl_idname = "OBJECT_PT_window_generator"
     bl_label = "Window Rail"
     bl_space_type = 'VIEW_3D'
